# Reuse a byte buffer for holding XML