@andrewkroh 2016-12-05
I should probably re-run all these benchmarks at the same time. They were taken at different times using vagrant (some while on battery power). The amount of bytes being allocated is smaller, but the batch processing times are a getting slower. Overall they _should_ be faster given the smaller allocs and shorter GC pauses.

Almost all of the time spent in reading event logs is in the [`EvtRender`](https://github.com/elastic/beats/blob/master/winlogbeat/sys/wineventlog/wineventlog_windows.go#L196) syscall. The second largest time consumer is XML unmarshalling (which is only like 3% of the total time). So there's not much left to optimize with the current architecture. There potentially could be some improvements with some architecture changes where `EvtRender` was called concurrently to process the batch of events.