@andrewkroh 2016-12-06
**Don't merge this yet.**

After repeatedly running the tests I can confirm that this PR is slower than what's in master.

I generated a CPU profile while running the benchmark and noticed that the `bytes.Buffer.Write()` was taking up a little more time than I expected for such a simple operation.

I wrote a really simple `io.Writer` wrapper around a []byte and retested with better results. I will work on updating the PR.

```
--- PASS: TestBenchmarkBatchReadSize (71.20s)
        bench_test.go:100: batch_size=10, total_events=30000, batch_time=5.340305ms, events_per_sec=1872.552223140813, b
ytes_alloced_per_event=14 kB, total_allocs=7295841
        bench_test.go:100: batch_size=100, total_events=30000, batch_time=52.326326ms, events_per_sec=1911.0839159623015
, bytes_alloced_per_event=14 kB, total_allocs=7264351
        bench_test.go:100: batch_size=500, total_events=25000, batch_time=278.975958ms, events_per_sec=1792.26913883382,
 bytes_alloced_per_event=14 kB, total_allocs=6050738
        bench_test.go:100: batch_size=1000, total_events=30000, batch_time=530.56368ms, events_per_sec=1884.787892001955
5, bytes_alloced_per_event=14 kB, total_allocs=7260423

--- PASS: TestBenchmarkBatchReadSize (69.01s)
        bench_test.go:100: batch_size=10, total_events=30000, batch_time=5.326304ms, events_per_sec=1877.474511406033, b
ytes_alloced_per_event=14 kB, total_allocs=7295807
        bench_test.go:100: batch_size=100, total_events=30000, batch_time=50.056196ms, events_per_sec=1997.7546835560574
, bytes_alloced_per_event=14 kB, total_allocs=7264316
        bench_test.go:100: batch_size=500, total_events=25000, batch_time=273.635652ms, events_per_sec=1827.247276973981
3, bytes_alloced_per_event=14 kB, total_allocs=6050700
        bench_test.go:100: batch_size=1000, total_events=30000, batch_time=511.429253ms, events_per_sec=1955.30465677136
36, bytes_alloced_per_event=14 kB, total_allocs=7260395
PASS
```